# msvc-specification/src/main/resources/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Swagger Microservices Bank
  description: Microservices Bank
  termsOfService: http://swagger.io/terms/
  contact:
    name: Swagger API Team
    email: apiteam@swagger.io
    url: http://swagger.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://clientstore.swagger.io/api
paths:
...
  /finantial/account/{id}/movement:
    post:
      tags:
        - FINANTIAL
      summary: Deposit or withdrawal operations
      description: Deposit or withdrawal operations.  Duplicates are not allowed
      operationId: movement
      parameters:
        - name: id
          in: path
          description: ID of account to movement
          required: true
          schema:
            type: string
      requestBody:
        description: Deposit or withdrawal operations
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovementDto'
      responses:
        '200':
          description: Account response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
...
components:
  schemas:
...
    MovementDto:
      allOf:
        - type: object
          required:
            - id
          properties:
            id:
              type: "string"
            type:
              type: "string"
            amount:
              type: "number"
              format: "double"
    AccountDto:
      allOf:
        - type: object
          required:
            - id
          properties:
            balance:
              type: "number"
              format: "double"
...
    ErrorDto:
      type: object
      required:
        - message
      properties:
        message:
          type: string